refactor: move spark-api files to a subdir
