+++ spark-api/index.js
