+++ spark-api/lib/instrument.js
