+++ spark-api/lib/round-tracker.js
