+++ spark-api/lib/validate.js
