+++ spark-api/test/round-tracker.test.js
@@ -6,7 +6,7 @@ import {
   mapCurrentMeridianRoundToSparkRound,
   startRoundTracker
 } from '../lib/round-tracker.js'
-import { migrate } from '../migrations/index.js'
+import { migrate } from '../../migrations/index.js'
 import { assertApproximately } from './test-helpers.js'
 import { createMeridianContract } from '../lib/ie-contract.js'
 import { afterEach, beforeEach } from 'mocha'
