+++ spark-api/test/test-helpers.js
