--- spark-publish/package.json
@@ -1,33 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "start": "node bin/spark-publish.js",
-    "test": "standard && mocha"
-  },
-  "dependencies": {
-    "@filecoin-station/spark-impact-evaluator": "^1.0.0",
-    "@glif/filecoin-address": "^3.0.5",
-    "@influxdata/influxdb-client": "^1.33.2",
-    "@ipld/car": "^5.3.0",
-    "@sentry/node": "^8.4.0",
-    "@ucanto/core": "^10.0.1",
-    "@ucanto/principal": "^9.0.1",
-    "@web3-storage/access": "^18.3.2",
-    "@web3-storage/w3up-client": "^13.1.1",
-    "ethers": "^6.12.1",
-    "p-retry": "^6.2.0",
-    "p-timeout": "^6.1.2",
-    "pg": "^8.11.5"
-  },
-  "devDependencies": {
-    "@flydotio/dockerfile": "^0.5.7",
-    "mocha": "^10.4.0",
-    "multiformats": "^13.1.0",
-    "standard": "^17.1.0"
-  },
-  "standard": {
-    "env": [
-      "mocha"
-    ]
-  }
-}
